Split ZPlainObject from ZObject
ZObject is now supposed to be the base class of all our runtime values,
including ZCharacter etc. (in practice, that’s not done yet, they will
be migrated in separate commits). ZPlainObject is the concrete class for
runtime objects of no more specific type, to be created by nodes like
ZObjectLiteralNode.
