+++ language/src/main/java/de/lucaswerkmeister/graaleneyj/nodes/ZPairNode.java
@@ -26,7 +26,7 @@ public abstract class ZPairNode extends Node {
 			@CachedLibrary(limit = "3") DynamicObjectLibrary putType,
 			@CachedLibrary(limit = "3") DynamicObjectLibrary putFirst,
 			@CachedLibrary(limit = "3") DynamicObjectLibrary putSecond) {
-		DynamicObject pair = context.makeObject(Map.of());
+		DynamicObject pair = context.makePlainObject(Map.of());
 		putType.put(pair, ZConstants.ZOBJECT_TYPE, new ZReference(ZConstants.PAIR));
 		putFirst.put(pair, ZConstants.PAIR_FIRST, first);
 		putSecond.put(pair, ZConstants.PAIR_SECOND, second);
